+++ src/constants/domain.ts
@@ -1,0 +1,5 @@
+export const FREE_DELIVERY_THRESHOLD = 100000;
+export const DELIVERY_FEE = 3000;
+export const EXTRA_REMOTE_FEE = 3000;
+
+export const MAX_COUPON_COUNT = 2;
